Copy over r37497 from the trunk:
  ------------------------------------------------------------------------
  r37497 | alagoutte | 2011-05-31 14:41:43 -0700 (Tue, 31 May 2011) | 8 lines
  Changed paths:
     M /trunk/epan/dissectors/packet-eapol.c
     M /trunk/epan/dissectors/packet-ieee80211.c
     M /trunk/epan/dissectors/packet-ieee80211.h

  From Mark Phillips via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5830
  802.11 Tag WME Qos Info field is not decoded

  From me :
  * Update from Trunk
  * Make item fiterable
  ------------------------------------------------------------------------
  r37505 | tuexen | 2011-06-01 02:26:33 -0700 (Wed, 01 Jun 2011) | 3 lines
  Changed paths:
     M /trunk/gtk/main.c

  Improve handling of command line arg errors (for example "wireshark -f").
  From Irene Ruengeler.
  ------------------------------------------------------------------------
  r37511 | cmaynard | 2011-06-01 11:34:41 -0700 (Wed, 01 Jun 2011) | 2 lines
  Changed paths:
     M /trunk/epan/dissectors/packet-dtls.c

  From Michael Chen via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5863, with some additional enhancements by me:  Add dtls heuristics.
  ------------------------------------------------------------------------
  r37521 | wmeier | 2011-06-02 07:22:19 -0700 (Thu, 02 Jun 2011) | 3 lines
  Changed paths:
     M /trunk/epan/dissectors/packet-opensafety.c

  Fix Bug #5984: Infinite loop caused by fcn call arg conversion: guint->guint8
  https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5984
  ------------------------------------------------------------------------


Copy over r37486 with the help of Alexis:

  ------------------------------------------------------------------------
  r37486 | alagoutte | 2011-05-31 12:30:53 -0700 (Tue, 31 May 2011) | 9 lines
  Changed paths:
     M /trunk/epan/dissectors/packet-ieee80211.c

  Enhance Vendor Specific IE (227)
  Specifically WPA/WME sub dissector
  * Rework from scratch VS WPA/WME dissector
  * Replace proto_tree_add_text/string (ugly hf_ieee80211_tag_interpretation header field...) by proto_tree_add_item
  * Make item filterable
  * and other stuff...

  Based on WMM_Specification_1-1
  ------------------------------------------------------------------------

svn path=/trunk-1.6/; revision=37522

+++ epan/dissectors/packet-opensafety.c
@@ -400,7 +400,7 @@ static guint16 crc16_opensafety(guint32 len, guint8 * pBuffer, guint16 initCRC);
 
 static guint stringToBytes( const char * string, guint8 * pBuffer, guint32 length );
 
-static guint8 findSafetyFrame ( guint8 * pBuffer, guint32 length, guint8 u_Offset, guint *u_frameOffset, guint *u_frameLength );
+static guint8 findSafetyFrame ( guint8 * pBuffer, guint32 length, guint u_Offset, guint *u_frameOffset, guint *u_frameLength );
 
 static guint stringToBytes( const char * stringToBytes, guint8 * pBuffer, guint32 length )
 {
@@ -613,7 +613,7 @@ static guint8 crc8_opensafety(guint32 len, guint8 * pBuffer, guint8 initCRC)
     return crc;
 }
 
-static guint8 findSafetyFrame ( guint8 * pBuffer, guint32 length, guint8 u_Offset, guint *u_frameOffset, guint *u_frameLength )
+static guint8 findSafetyFrame ( guint8 * pBuffer, guint32 length, guint u_Offset, guint *u_frameOffset, guint *u_frameLength )
 {
     guint n;
     guint16 crc, calcCrc;
